requirements.txt
Django>=4.2.0
pandas>=1.5.0
python-dotenv>=1.0.0
Pillow>=9.0.0
resend>=0.6.0
requests>=2.31.0
mjml>=0.11.0
premailer>=3.10.0
beautifulsoup4>=4.12.0
psycopg2-binary>=2.9.0
gunicorn>=21.0.0
